Added Shapiro-Wilk test
I added a Shapiro-Wilk test to the function that tests if the data provided is normally distributed. This has to be checked for each peptide separately as only if both - the input data and the calculated model - follow normal distribution, a Pearson correlation should be calculated. In the output the Shaprio-Wilk p-value will be reported alongside an additional column that contains logicals to indicate if the data follows normal distribution (p-value < 0.05).

I did not add a filtering option as I think users should decide if they want to filter the data based on that p-value or not.