Quick fix fetch_uniprot
UniProt recently updated their website and changed the way it is accessed programmatically. I changed the URL of the UniProt website to the legacy version as a quick fix. Still need to change the fetch_uniprot function completely though, to ensure that it will work with the newest UniProt release.